:recycle: refactor runtime calls into shared module
