--- osc_sdk_python/runtime/async_/__init__.py
@@ -1,5 +1,0 @@
-"""Asynchronous runtime implementation."""
-
-from .call import AsyncCall
-
-__all__ = ["AsyncCall"]
