--- osc_sdk_python/runtime/sync/__init__.py
@@ -1,5 +1,0 @@
-"""Synchronous runtime implementation."""
-
-from .call import Call
-
-__all__ = ["Call"]
